+++ gm2m/__init__.py
@@ -1,5 +1,6 @@
 from .version import __version__, __version_info__
 
+from . import compat
 from .fields import GM2MField
 
 default_app_config = 'gm2m.apps.GM2MConfig'