+++ gm2m/models.py
@@ -52,15 +52,15 @@ def create_gm2m_intermediary_model(field, klass):
         'Meta': meta,
         '__module__': klass.__module__,
         SRC_ATTNAME: models.ForeignKey(klass,
-                                       on_delete=field.rel.on_delete_src,
-                                       db_constraint=field.rel.db_constraint),
+                                       on_delete=field.remote_field.on_delete_src,
+                                       db_constraint=field.remote_field.db_constraint),
         CT_ATTNAME: models.ForeignKey(ct.ContentType,
-                                      db_constraint=field.rel.db_constraint),
+                                      db_constraint=field.remote_field.db_constraint),
         FK_ATTNAME: models.CharField(max_length=fk_maxlength),
         TGT_ATTNAME: ct.GenericForeignKey(
             ct_field=CT_ATTNAME,
             fk_field=FK_ATTNAME,
-            for_concrete_model=field.rel.for_concrete_model,
+            for_concrete_model=field.remote_field.for_concrete_model,
         ),
     })