+++ gm2m/query.py
@@ -36,7 +36,11 @@ class GM2MTgtQuerySet(query.QuerySet):
             rel_prefetching = False
 
         ct_attrs = defaultdict(lambda: defaultdict(lambda: []))
+        objects = {}
+        ordered_ct_attrs = []
+
         field_names = self.model._meta._field_names
+        fk_field = self.model._meta.get_field(field_names['tgt_fk'])
 
         extra_select = list(self.query.extra_select)
 
@@ -44,30 +48,43 @@ class GM2MTgtQuerySet(query.QuerySet):
                                    field_names['tgt_fk'],
                                    *extra_select):
             ct = vl[0]
-            pk = vl[1]
+            pk = fk_field.to_python(vl[1])
             ct_attrs[ct][pk].append(vl[2:])
+            ordered_ct_attrs.append((ct, pk))
 
         for ct, attrs in six.iteritems(ct_attrs):
             for pk, obj in six.iteritems(
                 ct_classes.ContentType.objects.get_for_id(ct).model_class()
                                       ._default_manager.in_bulk(attrs.keys())):
 
+                pk = fk_field.to_python(pk)
+
                 # we store the through model id in case we are in the process
                 # of fetching related objects
                 for i, k in enumerate(extra_select):
                     e_list = []
-                    for e in attrs[str(pk)]:
+                    for e in attrs[pk]:
                         e_list.append(e[i])
                     setattr(obj, k, e_list)
 
                 if rel_prefetching:
                     # when prefetching related objects, one must yield one
                     # object per through model instance
-                    for __ in attrs[str(pk)]:
-                        yield obj
+                    for __ in attrs[pk]:
+                        if self.ordered:
+                            objects[(ct, pk)] = obj
+                        else:
+                            yield obj
                     continue
 
-                yield obj
+                if self.ordered:
+                    objects[(ct, pk)] = obj
+                else:
+                    yield obj
+
+        if self.ordered:
+            for ct, pk in ordered_ct_attrs:
+                yield objects[(ct, pk)]
 
     def filter(self, *args, **kwargs):
         model = kwargs.pop('Model', None)